Feat: Add architecture components
-LiveData for lifecycle aware observing.

-ViewModel to manage UI data.

-Retrofit for a type-safe HTTP client.
